multiattach is capable after all
Turns out we're in fact able to do everything we want in multiattach
mode. The only issue being that the multiattach keyword should not be
used in the storageattach command.

Some better documentation or error reporting would have been nice.
Stumbled upon the following golden tip after having finished up all of
my other tests and tweaks (immutable -> multiattach, clonehd fixup).

https://www.virtualbox.org/ticket/11700

This makes the --clonehd feature obsolete (after having existed for
just an hour or so) but we can keep it in there just in case somebody
ever finds a use case for it.