workaround for disappearing harddisks and their type
The previous changes regarding multiattach have improved the way cloning
works. However, as it turns, out this is not enough. When a harddisk is
not attached to a Virtual Machine it will disappear after just a couple
of seconds.
In other words, our cloning would only work if you vmcloak-clone *right*
after having created a bird image with vmcloak. As this is not
necessarily the case this commit "fixes" this by setting the harddisk
type right before attaching it.
Now to hope this holds in all environments and will not give us any more
race condition issues.