introduce new vmcloak commands and office dependency
The upcoming 0.3 version of VMCloak will feature a couple of new
commands. Mostly reinventing the wheel of course, but in a way that
improves the framework overall and allows easier extension later on.

Features the vmcloak-modify tool which can be used to modify an existing
image to have it include whatever you like by doing manual changes.
Features the vmcloak-install tool which allows one to install known
dependencies. For now only the Office 2007 dependency is included.