introduce image modes and a snapshot table
After a snapshot has been made of an image, and thus the HDD has been
modified to become a multiattach device, it is no longer to update this
HDD. It is possible, however, to clone it and modify the clone.
Instruct people to do exactly so. An alternative will be to remove the
snapshots, later on.