enable versions for dependencies and set resolution
Currently doesn't default to any special resolution, but allows one to
either set a resolution when making a snapshot, or as a permanent option
by "installing" the resolution as a dependency with a version.
The latter can be achieved through the following (where the colon
indicates a version identifier - in this case used as resolution).

vmcloak install vmname resolution:1024x768