Support for multiple urls per exe, replace wget.
Replaced the system call to wget by a python downloader using urllib2.

You can now add multiple urls per exe as source. If a mirror is down, it tries the next one. Its backwards compatible with the old url field in exe. If the a list of urls is given, the depricated url field is ignored.